initial implementation for retrying messages in the event of failure
The main reason I wanted to do this was to properly handle the rate
limiting case (HTTP 429), but I haven't actually been able to
demonstrate that in practice, even when rapidly sending many messages
in a tight loop.

I decided to extend this not just to other HTTP exceptions, but to any
generic Exception when communicating to the server repeatedly in the
post_messages() loop. For now that is just implemented for
channel.send() (wrapped with send_msg()), but we should probably do so
for the threaded case (sent_message.create_thread()) and
created_thread.send()) as well, but that's not yet
implemented. Probably will *not* do for the call to
get_channel(), (just one for now, eventually one per channel), which
we should probably instead just do all up front and fail fast if
there's an issue.

For now there's no backoff, that will be come soon.

Also for now there's an artificially inserted 1 second delay between
sending messages to make this easier to debug. I found it difficult to
programattically cause failures, the easiest way is from the GUI for
the channel being tested, temporarily disable the "Send Messages"
permission via:

  <channel> -> right click -> Edit Channel -> Permissions
  -> Advanced permissions -> <Bot user> -> Text Channel Permissions
  -> Send Messages -> [x] -> Save Changes