implement dry run
we don't need to support arbitrary args and kwargs for our subclass of
the Discord client

even supporting arbitrary kwargs is sort of overkill, since we're not
using them

but it's easy to keep that. ditch the args though b/c I want to make
for now verbose and dry_run kwargs with defaults (and maybe will add
more later).