add optional --server argument
This is the name of the Discord server, which at low level is (I
think) synonymous with a guild. This is only needed in
the (presumably) uncommon case in which your bot is a member of
multiple servers.

The initialization of the discord.Client gives us some data that we
can take advantage of via attributes, to avoid HTTP API calls from
calling methods. This includes client.guilds, rather than having to
call fetch_guilds(). (Previously we were doing neither, and were just
ignoring the possibility of multiple guilds, but this is cleaner.)

Similarly, once we have the Guild object, we can use guild.channels
rather than having to call client.get_all_channels(), like we were
before. But rather than using all of the channels and then logging a
warning if the found channel is not a TextChannel, limit the search to
just guild.text_channels. This is a text import tool, voice is not
supported.

In the (unlikely, if not even impossible) event of multiple guilds
with the same name, just give up, rather than arbitrarily picking the
first. And now do the same for multiple channels with the same
name. (The latter is a change from before.)

For now, still fail if a destination channel does not exist. All of
this is a prelude to being able to optionally create text channels in
this case. create_text_channel() is a method of Guild, which is what
prompted the introduction of explicitly handling configuration at the
guild level.