Add the option to create missing Discord destination channels
The default is to still fail if a channel does not exist.
Create channels in this case using the new `--create` command line
option.

To get the new channel(s) properly placed in the GUI (in the category
"Text Channels", there's another round of object searching to find the
appropriate category (see
https://discordpy.readthedocs.io/en/latest/api.html#discord.Guild.categories)