Better processing of name within Slack messages
Previously, always used 'real_name' in 'user_profile'. A user profile
is not always present (bot case).

Now, this is a searchlist, abstracted in get_name():
* 'display_name' in 'user_profile'
* 'real_name' in 'user_profile'
* 'user', with leading 'U' stripped if present
* '???' if all else fails

Additionally, use 'type' of 'message' as the trigger for a
message to parse. Was previously using the existence of the keys
'user_profile' (which was wrong in the bot case) and 'ts' and
'text'.

This fixes https://github.com/richfromm/slack2discord/issues/13

According to the docs at:
https://slack.com/help/articles/220556107-How-to-read-Slack-data-exports#how-to-read-messages

'ts' and 'text' should always be present (although the latter
might be the empty string). Regardless, defensively code in the
event that they might not be.  If 'ts' is missing, log a warning
and skip the message.  If 'text' is missing, default to the empty
string.