add support for handling links
slack calls these attachments
discord calls these embeds

for a slack message with an attachment, get the relevant attributes
when parsing the slack export, and add them to a new list of
MessageLink objects (this is a new class) within the ParsedMessage
object.

the terminology within MessageLink comes from slack. translate to
discord terminology when calling get_discord_send_kwargs() to send the
message to discord.

note that discord has a limit of 10 embed's per message. parse all
attachments present in a slack export, but only send the first 10 to
discord as embeds.

slack URL's are escaped such that '/' is transformed to '\/' in the
export. undo this in appropriate fields when populating the
ParsedMessage.