add fetching of files from slack
based on what was parsed in previous step

still not uploading yet to discord

add a new optional arg, --downloads_dir. if not present, defaults to a
uniquely named timestamp dir (second accuracy) in a subdir of a
'downloads' subdir relative to the slack2discord.py script. directory
path will be created if necessary.

downloads are all sequential. could be made faster with
asyncio (e.g. aiohttp), but not worth the added complexity. at least
not for now.

downloads are not streamed in chunks, meaning that the entire file has
to fit in memory. it is assumed that the files are in practice small
enough that it's not worth the added complexity. again, at least not
for now.

very beginning of using mypy typing hints, at least for some of the
new code. still need to go back and add for existing code. (future
work.)