actually upload the files to discord
this is done by calling Message.add_files() on the Message object
returned from previously sending the message (to either a channel or a
thread)

wrap this in the method DiscordClient.add_files_to_message(), which is
decorated like the other methods making actual Discord HTTP API calls
with @discord_retry

add a new helper method ParsedMessage.get_discrod_add_file_args() to
help prepare the args to pass to Discord to add the files

all of the files are added together in a single API call, as opposed
to iterating through the list of files and individually adding each
one. my logic was that doing it together would be more efficient, but
I wonder about atomicity and idempotency in the event of
failure. sadly, this is difficult to test, and I don't think it's
documented.