get `mypy slack2discord.py` to pass
our virtualenv needs to additionally install
* pip install mypy
* pip install types-decorator
* pip install types-requests

this should definitely noted, and maybe even explicitly supported
somehow, maybe it's time for a requirements.txt file (and/or a
separate requirements-dev.txt file), and dropping the reference
to (and support of) pyv.

note that you should **not** `pip install discord.py-stubs`. the last
version of this is 1.7.3, and it must be consistent with
discord.py (currently 2.0.1), and doing so forces a downgrade. and
discord.utils.setup_logging (which we use in slack2discord.py) doesn't
exist in 1.7.3

but it appears to be no longer needed, as discord.py now has mypy
typing hints as of 2.0.0. see:

https://github.com/bryanforbes/discord.py-stubs/issues/171
https://github.com/Rapptz/discord.py/issues/5935
https://github.com/Rapptz/discord.py/issues/7342