Wrap the downloads in a progress bar with tqdm
As suggested in the issue
This is pretty coool

Decrease logging so that the progress bar in the non-verbose case
doesn't add a huge amount of clutter. Log skips at debug, not info.

And the downloads dir already existing is now an expected use case, so
log that at info, not warning.