add more mypy typing
This is getting pretty awkward in some cases, esp. when parsing JSON
in parser.py. For some hints, see:
https://stackoverflow.com/questions/51291722/define-a-jsonable-type-using-mypy-pep-526

I'm also wondering if this is maybe a sign that we should be using
classes better, and not so many dicts (e.g. `users`, `channel_map`,
and `parsed_messages` in `SlackParser`), some of which are deeply
nested just a bit too much.

This branch was rebased off of the edits from:
https://github.com/richfromm/slack2discord/pull/31

For that, we also now need to install stubs for tqdm:
* pip install tqdm-stubs

For some reason that I don't understand, the current state is
that the code passes mypy every other run.  In between, there is
the following error reported:
```
mypy slack2discord.py
/Users/rich/.virtualenvs/slack2discord/lib/python3.10/site-packages/discord/app_commands/commands.py:72: error: Cannot find implementation or library stub for module named "discord.ext"
/Users/rich/.virtualenvs/slack2discord/lib/python3.10/site-packages/discord/app_commands/commands.py:72: note: See https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-imports
Found 1 error in 1 file (checked 1 source file)
```

Apparently this was true even before the rebasing, as of:
https://github.com/richfromm/slack2discord/commit/b5652eaec52526145b880319a116fa5617579059

Ignore this for the immediate time being.