Finish applying mypy typing hints
This all turned out to be a bit more difficult than I had hoped.

There are more uses of Optional and cast() required than I'd
like. Some of this is I think do to the use of dry_run. There are
instances in which objects might be None for a dry run only, but not
for a real execution. And then those cases propogate. I was hoping
that overloading might help here, but I'm thinking maybe not:

https://mypy.readthedocs.io/en/stable/more_types.html#function-overloading

Our use of complex data structures, in many (but not all) cases
related to the import of JSON from the Slack export, is also
complicating things.

For some ideas, see:
https://stackoverflow.com/questions/51291722/define-a-jsonable-type-using-mypy-pep-526

We could probably simplify some if this using `NewType`, which is like
a typedef. I will explore this next:
https://mypy.readthedocs.io/en/stable/more_types.html#newtypes

But this also might somewhat be a sign that we should be defining more
classes rather than using complicated multi-level nested dict's.

There were a number of times where mypy complained b/c I had a
method in my `DiscordClient` with the same name as a method in the
superclass `discord.Client`, but with a different signature. I
resolved this by renaming my functions:
   DiscordClient.get_guild_maybe_by_name() vs. discord.Client.get_guild()
   DiscordClient.get_channel_by_name() vs. discord.Client.get_channel()
See https://stackoverflow.com/questions/51003146/python-3-6-signature-of-method-incompatible-with-super-type-class

mypy is confused in various cases by my use of `*args` and `**kwargs`. I
might be able to improve upon this, but it's not the highest priority right
now. For now just don't type some of these, and ignore some reported
errors. For possible future reference:
https://stackoverflow.com/questions/62574831/why-do-i-get-mypy-no-overload-variant-of-zip-matches-argument-error

There are now a few different ways in which mypy is confused by our
use of decorators. For now (and probably for ever) just ignore them.

AFAIK, an empty `return` should be the same as `return None`. However, in
all cases where this existed within an `if self.dry_run` condition, in
which the declared method return type was some `Optional[]`, mypy didn't
like this. So I just added the explicit `return None`.

The addition of mypy actually found a few legit errors in client.py:

* The `error_msg` in [new name] `get_channel_by_name()` referenced
  `channel_ids`, which does not exist (create the list now with a simple
  list comprehension)

* The handling of the `discord.RateLimited` exception in `discord_retry()`
  had a typo referencing the exception to obtain `retry_after`

There's an awful lot of `discord.Foo` references now, esp. in
client.py. Perhaps it's time to be more specific than the blanket
`import discord`

Include a mypy config file, for now using it to enable error
codes (even though allegedly `--show-error-codes` should be the
default behavior). This makes it easier to know how to insert comments
to ignore certain errors. See:
https://mypy.readthedocs.io/en/stable/error_codes.html