Validate Discord channel names
The following criteria must be met:
* Names must contain only alphanumeric, dash (-), and/or underscore (_) characters
* Length must be between 1 and 100 chars (inclusive)
* Multiple dashes in a row (e.g. --) are not permitted

The actual reality of Discord channel creation is a bit more complex. Empirically, there
are numerous (undocumented?) cases where an input channel name does **not** fail to create
a channel, but the channel created does not have precisely the same name as the input. For
example:
* Multiple dashes in a row are converted into a single dash
* Space and tilde are converted to dash
* Tab and newline are removed
* All other symbols are removed

For our purposes, this is too complex to deal with, and we will fail to validate any input
that does not result in creating a channel with the same name as the input.

https://github.com/richfromm/slack2discord/issues/24